requirements.txt
streamlit
kagglehub
pandas
numpy
matplotlib
SentenceTransformer
os
NearestNeighbors
plotly.express
umap